[ADD] Extra visualization example

+++ autoPyTorch/api/base_task.py
@@ -576,6 +576,7 @@ class BaseTask:
         assert self._dask_client is not None
 
         self._logger.info("Starting to create traditional classifier predictions.")
+        starttime = time.time()
 
         # Initialise run history for the traditional classifiers
         run_history = RunHistory()
@@ -649,6 +650,7 @@ class BaseTask:
                         origin = additional_info['configuration_origin']
                         run_history.add(config=configuration, cost=cost,
                                         time=runtime, status=status, seed=self.seed,
+                                        starttime=starttime, endtime=starttime + runtime,
                                         origin=origin)
                     else:
                         if additional_info.get('exitcode') == -6:
@@ -1235,7 +1237,8 @@ class BaseTask:
         # When a multiprocessing work is done, the
         # objects are deleted. We don't want to delete run areas
         # until the estimator is deleted
-        self._backend.context.delete_directories(force=False)
+        if hasattr(self, '_backend'):
+            self._backend.context.delete_directories(force=False)
 
     @typing.no_type_check
     def get_incumbent_results(
